+++ node/modules/index.js
@@ -1,4 +1,5 @@
 // for exporting/making modules public to app.js
 // kind of like internal API
 
-export * from './controller.js';
+export * as controller from './controller.js';
+export * from './utils.js';
